feat: Implement incremental training functionality
